Export to pdf

+++ backend/controllers/notes.js
@@ -1,22 +1,23 @@
 const notesRouter = require('express').Router()
 const Note = require('../models/note')
 const User = require('../models/user')
-const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
+const { ObjectId } = require('mongodb')
+const { getIo } = require('../socket')
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
+notesRouter.use(userExtractor)
 
-notesRouter.get('/', async (request, response) => {
-  const notes = await Note
-    .find({}).populate('user', { username: 1, name: 1 })
+// get all notes for a user
+notesRouter.get('/:userId', async (request, response) => {
+  const { userId } = request.params
+  console.log('userId:', userId)
+  const user = new ObjectId(userId)
+  const notes = await Note.find({ user })
+  console.log('notes:', notes)
   response.json(notes)
 })
 
+// get a single note
 notesRouter.get('/:id', async (request, response) => {
   const note = await Note.findById(request.params.id)
 
@@ -30,37 +31,41 @@ notesRouter.get('/:id', async (request, response) => {
 notesRouter.post('/', async (request, response) => {
   const body = request.body
 
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-  if(!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
-  }
-
-  const user = await User.findById(decodedToken.id)
-
   const note = new Note({
+    title: body.title,
     content: body.content,
     important: body.important === undefined ? false : body.important,
-    user: user.id
+    user: request.user._id,
+    id: body.id
   })
 
   const savedNote = await note.save()
-  user.notes = user.notes.concat(savedNote._id)
-  await user.save()
+  getIo().emit('noteAdded', savedNote)
 
   response.status(201).json(savedNote)
 })
 
 notesRouter.delete('/:id', async (request, response) => {
+  const note = await Note.findById(request.params.id)
+  if (!note) {
+    return response.status(404).json({ error: 'Note not found' })
+  }
+
+  const user = await User.findById(note.user)
+  user.notes = user.notes.filter(n => n.toString() !== note.id.toString()) // remove note from user
+  await user.save()
+
   await Note.findByIdAndDelete(request.params.id)
+  getIo().emit('noteDeleted', request.params.id)
   response.status(204).end()
 })
 
 notesRouter.put('/:id', async (request, response) => {
-  const { content, important } = request.body
+  const { title, content, important } = request.body
 
   const updatedNote = await Note.findByIdAndUpdate(
     request.params.id,
-    { content, important },
+    { title, content, important },
     { new: true, runValidators: true, context: 'query' }
   )
